+++ src/products/ConsumerElectronics.java
@@ -7,13 +7,13 @@ public abstract class ConsumerElectronics extends Product {
     private Double screenSize;
     private int storageSize;
 
-    public ConsumerElectronics(String id, String brand, String model, Double price, String type, Double height, Double width, Double depth, int weight, Double reviewPoint, Double screenSize, int storageSize) {
+    public ConsumerElectronics(String id, String brand, String model, Double price, String type, Double height, Double width, Double depth, Double weight, Double reviewPoint, Double screenSize, int storageSize) {
         super(id, brand, model, price, "Consumer Electronics", type , height, width, depth, weight, reviewPoint);
         this.screenSize = screenSize;
         this.storageSize = storageSize;
     }
 
-    public ConsumerElectronics(String brand, String model, Double price, String type, Double height, Double width, Double depth, int weight, Double screenSize, int storageSize) {
+    public ConsumerElectronics(String brand, String model, Double price, String type, Double height, Double width, Double depth, Double weight, Double screenSize, int storageSize) {
         super(brand, model, price, "Consumer Electronics", type , height, width, depth, weight);
         this.screenSize = screenSize;
         this.storageSize = storageSize;
