+++ src/products/MajorAppliance.java
@@ -7,13 +7,13 @@ public abstract class MajorAppliance extends Product {
     private int capacity;
     private String energyEfficiency;
 
-    public MajorAppliance(String id, String brand, String model, Double price, String type, Double height, Double width, Double depth, int weight, Double reviewPoint, int capacity, String energyEfficiency) {
+    public MajorAppliance(String id, String brand, String model, Double price, String type, Double height, Double width, Double depth, Double weight, Double reviewPoint, int capacity, String energyEfficiency) {
         super(id, brand, model, price, "Major Appliance", type, height, width, depth, weight, reviewPoint);
         this.capacity = capacity;
         this.energyEfficiency = energyEfficiency;
     }
 
-    public MajorAppliance(String brand, String model, Double price, String type, Double height, Double width, Double depth, int weight, int capacity, String energyEfficiency) {
+    public MajorAppliance(String brand, String model, Double price, String type, Double height, Double width, Double depth, Double weight, int capacity, String energyEfficiency) {
         super(brand, model, price, "Major Appliance", type, height, width, depth, weight);
         this.capacity = capacity;
         this.energyEfficiency = energyEfficiency;
