+++ src/products/Motorcycle.java
@@ -8,13 +8,13 @@ public class Motorcycle extends Vehicle {
     private boolean windshield;
     private boolean carrierBox;
 
-    public Motorcycle(String id, String brand, String model, Double price, Double height, Double width, Double depth, int weight, Double reviewPoint, int power, String fuelType, boolean windshield, boolean carrierBox) {
+    public Motorcycle(String id, String brand, String model, Double price, Double height, Double width, Double depth, Double weight, Double reviewPoint, int power, String fuelType, boolean windshield, boolean carrierBox) {
         super(id, brand, model, price, "Motorcycle", height, width, depth, weight, reviewPoint, power, fuelType);
         this.windshield = windshield;
         this.carrierBox = carrierBox;
     }
 
-    public Motorcycle(String brand, String model, Double price, Double height, Double width, Double depth, int weight, int power, String fuelType, boolean windshield, boolean carrierBox) {
+    public Motorcycle(String brand, String model, Double price, Double height, Double width, Double depth, Double weight, int power, String fuelType, boolean windshield, boolean carrierBox) {
         super(brand, model, price, "Motorcycle", height, width, depth, weight, power, fuelType);
         this.windshield = windshield;
         this.carrierBox = carrierBox;
